Sync pinner on every pin operation by default
This PR builds on PR #13 to sync the pinner on every pin operation. This PR does the following that #13 does not:

- Sync's dag service separately from pin data
- Does not release and immediately reacquire lock to sync, syncs while still holding pinner lock.
- Syncs only pin data for most operations

In addition to sync of pin data, this PR also revises how indexes are rebuilt.  Instead of reading through all pins and loading all previous index data, now only a single pass through the pins is needed to rebuild missing indexes resulting from incomplete add or delete operations. This is operationally much simpler, but also does not require storing entire pin sets or index sets in memory as did the previous solution.